Added support for returning a full jid
The previous implementation only returned the bare JID. I needed a full jit to pass to pass it to my client. Without the full jid, my pubsub was not working. Also fixed a couple issues to git it to work in my environment. Updated the spec for the new functionality.